Rename SplitButton.ts to SplitMenuButton.ts
