+++ packages/experiments/src/index.ts
@@ -11,7 +11,7 @@ export * from './PersonaCoin';
 export * from './Shimmer';
 export * from './Sidebar';
 export * from './Signals';
-export * from './SplitButton';
+export * from './SplitMenuButton';
 export * from './StaticList';
 export * from './Text';
 export * from './Tile';
